+++ src/flowcean/sklearn/__init__.py
@@ -1,3 +1,4 @@
+from .adaboost_classifier import AdaBoost
 from .linear_regression import LinearRegression
 from .metrics.classification import (
     Accuracy,
@@ -19,6 +20,7 @@ from .regression_tree import RegressionTree
 
 __all__ = [
     "Accuracy",
+    "AdaBoost",
     "ClassificationReport",
     "FBetaScore",
     "LinearRegression",